Excluded images do not remain selected in the UI after reprocessing

+++ uasdm-server/src/main/java-gen/stub/gov/geoplatform/uasdm/bus/Imagery.java
@@ -157,7 +157,7 @@ public class Imagery extends ImageryBase implements ImageryComponent, ImageryIF
   }
 
   @Override
-  public SiteObjectsResultSet getSiteObjects(String folder, Integer pageNumber, Integer pageSize)
+  public SiteObjectsResultSet getSiteObjects(String folder, Long pageNumber, Long pageSize)
   {
     List<SiteObject> objects = new LinkedList<SiteObject>();
 
@@ -197,7 +197,7 @@ public class Imagery extends ImageryBase implements ImageryComponent, ImageryIF
   }
 
   @Override
-  protected SiteObjectsResultSet getSiteObjects(String folder, List<SiteObject> objects, Integer pageNumber, Integer pageSize)
+  protected SiteObjectsResultSet getSiteObjects(String folder, List<SiteObject> objects, Long pageNumber, Long pageSize)
   {
     if (!folder.equals(RAW) && ( pageNumber != null || pageSize != null ))
     {
